feature_be: modularization of server code
