+++ server/node-server-app/src/config.js
@@ -1,0 +1,18 @@
+// @todo Use .env file to fill these values
+
+const SERVER_PORT = 9006;
+
+const DB_PORT = 54321;
+const DB_USER = "";
+const DB_PASS = "";
+const DB_NAME = "";
+const DB_HOST = "localhost";
+
+export {
+    SERVER_PORT,
+    DB_PORT,
+    DB_USER,
+    DB_PASS,
+    DB_NAME,
+    DB_HOST,
+};
