+++ server/node-server-app/src/middlewares/corsMiddleware.js
@@ -1,0 +1,29 @@
+import cors from 'cors';
+
+function setupCorsMiddleware(app) {
+    // Environment-specific origins
+    const allowedOrigins = [
+        'http://localhost:5173',
+        'https://localhost:5173'
+    ];
+
+    const corsOptions = {
+        origin: function (origin, callback) {
+            // Allow requests with no origin (mobile apps, etc.)
+            if (!origin) return callback(null, true);
+
+            if (allowedOrigins.indexOf(origin) !== -1) {
+                callback(null, true);
+            } else {
+                callback(new Error('Not allowed by CORS'));
+            }
+        },
+        credentials: true,
+        methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+        allowedHeaders: ['Content-Type', 'Authorization']
+    };
+
+    app.use(cors(corsOptions));
+}
+
+export { setupCorsMiddleware };
