--- server/node-server-app/tests-ava/server.test.js
@@ -1,7 +1,0 @@
-import test from 'ava';
-
-import { SERVER_PORT } from '../src/server.js';
-
-test("server", t => {
-    t.deepEqual(SERVER_PORT, 9006);
-});
