+++ server/node-server-app/tests-jasmine/config.spec.js
@@ -1,6 +1,6 @@
-import { SERVER_PORT } from "../src/server.js";
+import { SERVER_PORT } from "../src/config.js";
 
-describe("server", () => {
+describe("config", () => {
     it("servers traffic on port 9006", () => {
         expect(SERVER_PORT).toEqual(9006);
     });
